Bash: Add functionality finding indentation type & size of file
- Why:
Had the need to find which kind of indentation that was used as
it is to be used when inserting code into existing files.

- New/Changes:
script.sh implements and tests functionality.

find_indent_type() prints what is the most common indentation type
based on counting the amount of rows starting with one or the
other.

find_indent_size() prints the amount of spaces used for one level
of indentation when the indentation type is spaces. It bases this
on finding all indentations using spaces and picking the smallest
indentation size used.

find_indent() uses both of the above and prints the indentation
size if the indentation type is spaces.

- Testing:
Runs the functionalities with the examples files:
* example_1
* example_2
* example_3